Remove Disk.request() method.
The request() was removed because it should be handled in a wrapper. This allows
to access the same disk file from different APIs (for example node-lkl and fatfs).
It is replaced with the `read`, `write`, `flush` and `discard` methods.

The automatic opening of disks was also removed: you have to call the Disk.open() method
and wait for its callback before reading / writing.

The Disk.close() method was added.